+++ repos/common/src/main/kotlin/dev/inmo/micro_utils/repos/DatabaseTransactions.kt
@@ -2,15 +2,22 @@ package dev.inmo.micro_utils.repos
 
 import android.database.sqlite.SQLiteDatabase
 import dev.inmo.micro_utils.coroutines.safely
+import kotlinx.coroutines.newSingleThreadContext
 import kotlinx.coroutines.withContext
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.coroutineContext
+
+object InTransaction: CoroutineContext.Element, CoroutineContext.Key<InTransaction> {
+    override val key: CoroutineContext.Key<InTransaction> = InTransaction
+}
 
 suspend fun <T> SQLiteDatabase.transaction(block: suspend SQLiteDatabase.() -> T): T {
-    return withContext(DatabaseCoroutineContext) {
-        when {
-            inTransaction() -> {
-                block()
-            }
-            else -> {
+    return when {
+        coroutineContext[InTransaction] == InTransaction -> {
+            block()
+        }
+        else -> {
+            withContext(InTransaction) {
                 beginTransaction()
                 safely(
                     {